+++ src/bin/cmds/publish.rs
@@ -1,7 +1,7 @@
 use super::{args, get};
 use clap::{App, ArgMatches, SubCommand};
 use elba::{
-    cli::backend,
+    cli::registry,
     util::{config::Config, errors::Res},
 };
 use failure::{format_err, ResultExt};
@@ -9,7 +9,7 @@ use std::env::current_dir;
 
 pub fn cli() -> App<'static, 'static> {
     SubCommand::with_name("publish")
-        .about("Publishes a package to a backend")
+        .about("Publishes a package to a registry")
         .arg(args::index())
 }
 
@@ -21,13 +21,13 @@ pub fn exec(c: &mut Config, args: &ArgMatches) -> Res<String> {
     let bck = get::index(c, args)?;
     let bck_text = bck.to_string();
     let bcx = get::build_ctx(c, args);
-    let ctx = backend::BackendCtx {
+    let ctx = registry::RegistryCtx {
         index: bck,
         data_dir: c.directories.data.clone(),
     };
 
     // TODO: We require verification of packages before publishing; maybe not?
-    backend::publish(&bcx, &ctx, &project, true)?;
+    registry::publish(&bcx, &ctx, &project, true)?;
 
     Ok(format!("published package to index {}", bck_text))
 }