+++ src/lib/cli/mod.rs
@@ -1,5 +1,5 @@
 //! Handlers for all of the command-line actions of the binary.
 
-pub mod backend;
 pub mod build;
 pub mod new;
+pub mod registry;